Initialized form and and menu modules
