+++ modules/menu.js
@@ -1,7 +1,4 @@
-import { saveBooksToStorage } from './storage.js';
-import addBook from './addBook.js';
-
-export function initializeMenu() {
+export default function initializeMenu() {
   const menuItems = document.querySelectorAll('.row-list li a');
   menuItems.forEach((menuItem) => {
     menuItem.addEventListener('click', (event) => {
@@ -28,27 +25,5 @@ export function initializeMenu() {
       const targetSection = document.getElementById(targetSectionId);
       targetSection.classList.add('active');
     });
-  });
-}
-
-export function initializeForm(collection) {
-  const form = document.querySelector('#add-book-section form');
-  form.addEventListener('submit', (event) => {
-    event.preventDefault();
-
-    const titleInput = document.getElementById('title');
-    const authorInput = document.getElementById('author');
-
-    const title = titleInput.value;
-    const author = authorInput.value;
-
-    if (title && author) {
-      addBook(collection, title, author);
-      saveBooksToStorage(collection.books);
-      collection.displayBooks();
-
-      titleInput.value = '';
-      authorInput.value = '';
-    }
   });
 }
